backend/requirements.txt
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
black==21.12b0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.1
django-cors-headers==3.11.0
django-filter==21.1
django-oauth-toolkit==1.6.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
entrypoints==0.3
executing==0.8.2
fonttools==4.28.5
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.1
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jwcrypto==1.0
kiwisolver==1.3.2
lxml==4.7.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
mysqlclient==2.1.0
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.7
numpy==1.22.1
oauthlib==3.1.1
packaging==21.3
pandas==1.4.0rc0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.3
ptyprocess==0.7.0
pure-eval==0.2.1
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.6
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0rc2
Send2Trash==1.8.1b0
setuptools-scm==6.3.2
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.1.4
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tomli==1.2.3
torch==1.10.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.13.3
zipp==3.7.0
